GearSwap v0.810 - Instant precast/midcast
