+++ addons/GearSwap/equip_processing.lua
@@ -24,166 +24,6 @@
 --(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 --SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-
-function equip_sets(swap_type,val1,val2,ind)
-	load_globals(ind)
-	if debugging >= 1 then windower.debug(swap_type..' enter') end
-	_global.current_event = swap_type
-	refresh_globals()
-	local cur_equip = get_gs_gear(items.equipment,swap_type)
-	
-	table.reassign(equip_order,default_equip_order)
-	table.reassign(equip_list,to_names_set(cur_equip,items.inventory))
-	
-	if debugging >= 2 then windower.add_to_chat(8,swap_type) end
-	if logging then
-		logit(logfile,'\n\n'..tostring(os.clock)..'(15) equip_sets: '..tostring(swap_type))
-		if val1 then
-			if val1.english then
-				logit(logfile,' : '..val1.english)
-			end
-		else
-			logit(logfile,' : nil-or-false')
-		end
-		if val2 then
-			if val2.type then	logit(logfile,' : '..val2.type)	end
-		else
-			logit(logfile,' : nil-or-false')
-		end
-	end
-
-	if swap_type == 'pretarget' then
-		_global.cast_delay = 0
-		_global.verify_equip = false
-		_global.force_send = false
-		if _settings.debug_mode then windower.add_to_chat(8,'GearSwap (Debug Mode): Pretarget '..tostring(val1.name)) end
-		if type(user_env.pretarget) == 'function' then user_env.pretarget(val1,val2) -- User defined function to determine the pretarget set
-		elseif user_env.pretarget then windower.add_to_chat(123,'GearSwap: pretarget() exists but is not a function') end
-		ind = mk_out_arr_entry(val1,{target_id=spell.target.id},nil)
-	elseif swap_type == 'precast' then
-		if _settings.debug_mode then windower.add_to_chat(8,'GearSwap (Debug Mode): Precast '..tostring(val1.name)) end
-		if type(user_env.precast) == 'function' then user_env.precast(val1,val2) -- User defined function to determine the precast set
-		elseif user_env.precast then windower.add_to_chat(123,'GearSwap: precast() exists but is not a function') end
-		if _global.verify_equip then
-			windower.send_command('@wait 1;lua i '.._addon.name..' sender '..ind)
-		end
-		_global.midaction = true
-	elseif swap_type == 'midcast' then
-		if _settings.debug_mode then windower.add_to_chat(8,'GearSwap (Debug Mode): Midcast '..tostring(val1.name)) end
-		user_env.midcast(val1,val2) -- User defined function to determine the midcast set
-	elseif swap_type == 'aftercast' then
-		if _settings.debug_mode then windower.add_to_chat(8,'GearSwap (Debug Mode): Aftercast '..tostring(val1.name)) end
-		if not val1 then val1 = {}
-			if debugging >= 2 then
-				windower.add_to_chat(8,'val1 error')
-			end
-		end
-		_global.midaction = false
-		user_env.aftercast(val1,val2) -- User defined function to determine the aftercast set
-	elseif swap_type == 'pet_midcast' then
-		if _settings.debug_mode then windower.add_to_chat(8,'GearSwap (Debug Mode): Pet Midcast '..tostring(val1.name)) end
-		user_env.pet_midcast(val1,val2) -- User defined function to determine the midcast set
-	elseif swap_type == 'pet_aftercast' then
-		if _settings.debug_mode then windower.add_to_chat(8,'GearSwap (Debug Mode): Pet Aftercast '..tostring(val1.name)) end
-		if val1 then
-		else
-			windower.add_to_chat(8,'---------------- VAL1 IS NIL ------------')
-		end
-		user_env.pet_aftercast(val1,val2) -- User defined function to determine the aftercast set
-	elseif swap_type == 'status_change' then
-		if _settings.debug_mode then windower.add_to_chat(8,'GearSwap (Debug Mode): Status Change '..tostring(val1)..' '..tostring(val2))
-			for i=1,#val2 do
-				windower.add_to_chat(8,'GearSwap (Debug Mode): '..i..' '..string.byte(val2[i]))
-			end
-		end
-		if type(user_env.status_change) == 'function' then user_env.status_change(val1,val2) -- User defined function to determine if sets should change following status change
-		elseif user_env.status_change then windower.add_to_chat(123,'GearSwap: status_change() exists but is not a function') end
-	elseif swap_type == 'buff_change' then
-		if _settings.debug_mode then windower.add_to_chat(8,'GearSwap (Debug Mode): Buff Change '..tostring(val1)..' '..tostring(val2)) end
-		if type(user_env.buff_change) == 'function' then user_env.buff_change(val1,val2) -- User defined function to determine if sets should change following buff change
-		elseif user_env.buff_change then windower.add_to_chat(123,'GearSwap: buff_change() exists but is not a function') end
-	elseif swap_type == 'equip_command' then
-		equip(val1)
-	elseif swap_type == 'self_command' then
-		if _settings.debug_mode then windower.add_to_chat(8,'GearSwap (Debug Mode): Self Command '..tostring(val1)) end
-		if type(user_env.self_command) == 'function' then user_env.self_command(val1)
-		elseif user_env.self_command then windower.add_to_chat(123,'GearSwap: self_command() exists but is not a function') end
-	elseif swap_type == 'delayed' then
-		equip(stored_equip_list)
-	end
-	
-	if player.race == 'Precomposed NPC' then
-		-- Short circuit the routine and get out if there's no swapping to be done because the user is a monster.
-		if swap_type == 'pretarget' then
-			command_send_check(out_arr[ind].storedtarget)
-		elseif swap_type == 'precast' then
-			packet_send_check(true,ind)
-		elseif swap_type == 'aftercast' then
-			d_out_arr_entry(val1,ind)
-		end
-		return
-	end
-	
-	
-	for i,v in pairs(short_slot_map) do
-		if equip_list[i] and (disable_table[v] or encumbrance_table[v]) then
-			not_sent_out_equip[i] = equip_list[i]
-		end
-	end
-	
-	local equip_next = to_id_set(items.inventory,equip_list) -- Translates the equip_list from the player (i=slot name, v=item name) into a table with i=slot id and v=inventory id.
-	equip_next = eliminate_redundant(cur_equip,equip_next) -- Eliminate the equip commands for items that are already equipped
-	
-	if _settings.show_swaps and table.length(equip_next)>0 then
-		local tempset = to_names_set(equip_next,items.inventory)
-		print_set(tempset,swap_type)
-	end
-	
-	local failure_reason
-	for i,v in pairs(player.buffs) do
-		if v==14 or v == 17 then
-			failure_reason = 'Charmed'
-		elseif v == 0 then
-			failure_reason = 'KOed'
-		end
-		if _settings.debug_mode and failure_reason then
-			windower.add_to_chat(8,'GearSwap (Debug Mode): Cannot change gear right now: '..failure_reason)
-		end
-	end
-	
-	
-	if not failure_reason then
-		if ind and _global.verify_equip then sent_out_equip.ind = ind end
-		for _,i in ipairs(equip_order) do
-			if debugging >= 3 and equip_next[i] then
-				local out_str = 'Order: '..tostring(_)..'  Slot ID: '..tostring(i)..'  Inv. ID: '..tostring(equip_next[i])
-				if equip_next[i] ~= 0 then
-					out_str = out_str..'  Item: '..tostring(r_items[items.inventory[equip_next[i]].id][language..'_log'])
-				else
-					out_str = out_str..'  Emptying slot'
-				end
-				windower.add_to_chat(8,'GearSwap (Debugging): '..out_str)
-			elseif equip_next[i] and not disable_table[i] and not encumbrance_table[i] then
-				windower.debug('attempting to set gear. Order: '..tostring(_)..'  Slot ID: '..tostring(i)..'  Inv. ID: '..tostring(equip_next[i]))
-				windower.ffxi.set_equip(equip_next[i],i)
-				sent_out_equip[i] = equip_next[i] -- re-make the equip_next table with the name sent_out_equip as the equipment is sent out.
-			end
-		end
-	elseif logging then
-		logit(logfile,'\n\n'..tostring(os.clock)..'(69) failure_reason: '..tostring(failure_reason))
-	end
-	if debugging >= 1 then windower.debug(swap_type..' exit') end
-	
-	cache_globals(ind)
-	if swap_type == 'pretarget' then
-		return command_send_check(ind)
-	elseif swap_type == 'precast' then
-		return packet_send_check(_global.force_send,ind)
-	elseif swap_type == 'aftercast' then
-		d_out_arr_entry(val1,ind)
-	end
-end
-
 function check_wearable(item_id)
 	if not item_id or item_id == 0 then -- 0 codes for an empty slot, but Arcon will probably make it nil at some point
 	elseif not r_items[item_id] then
@@ -388,16 +228,6 @@ function to_names_set(id_id,inventory)
 	end
 	
 	return equip_package
-end
-
-function unify_slots(equipment)
-	local unified = {}
-	for i,v in pairs(equipment) do
-		if default_slot_map[slot_map[i]] then
-			unified[default_slot_map[slot_map[i]]] = v
-		end
-	end
-	return unified
 end
 
 function get_gs_gear(cur_equip,swap_type)
