+++ addons/GearSwap/flow.lua
@@ -1,40 +1,204 @@
+--Copyright (c) 2013, Byrthnoth
+--All rights reserved.
+
+--Redistribution and use in source and binary forms, with or without
+--modification, are permitted provided that the following conditions are met:
+
+--    * Redistributions of source code must retain the above copyright
+--      notice, this list of conditions and the following disclaimer.
+--    * Redistributions in binary form must reproduce the above copyright
+--      notice, this list of conditions and the following disclaimer in the
+--      documentation and/or other materials provided with the distribution.
+--    * Neither the name of <addon name> nor the
+--      names of its contributors may be used to endorse or promote products
+--      derived from this software without specific prior written permission.
+
+--THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+--ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+--WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+--DISCLAIMED. IN NO EVENT SHALL <your name> BE LIABLE FOR ANY
+--DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+--(INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+--LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+--ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+--(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+--SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+
+
 -----------------------------------------------------------------------------------
---Name: sender(...)
---Desc: Triggers an outgoing action packet if verify_equip()'s exit conditions are
---      still unmet after a second.
+--Name: equip_sets(swap_type,val1,val2,ind)
+--Desc: General purpose equipment pipeline / user function caller. 
 --Args:
----- {...} - space delimited key for out_arr (hopefully)
+---- swap_type - Determines equip_sets' behavior in terms of which user function it
+--      attempts to call
+---- val1 - First argument to be passed to the user function
+---- val2 - Second argument to be passed to the user function
+---- ind - nil or index of out_arr
 -----------------------------------------------------------------------------------
---Returns:
----- none
+--Return (varies by swap type):
+---- pretarget : empty string to blank packet or full string
+---- Everything else : nil
 -----------------------------------------------------------------------------------
-function sender(...)
-	local inde = table.concat({...},' ')
-	if not action_sent then
-		if debugging >= 1 or _settings.debug_mode then windower.add_to_chat(8,'GearSwap (Debug Mode): Had to force the command to send. Exit conditions went unmet.') end
-		packet_send_check(true,inde)
-		sent_out_equip = {}
+function equip_sets(swap_type,val1,val2,ind)
+	load_globals(ind)
+	if debugging >= 1 then windower.debug(swap_type..' enter') end
+	_global.current_event = swap_type
+	refresh_globals()
+	local cur_equip = get_gs_gear(items.equipment,swap_type)
+	
+	table.reassign(equip_order,default_equip_order)
+	table.reassign(equip_list,to_names_set(cur_equip,items.inventory))
+	
+	if debugging >= 2 then windower.add_to_chat(8,swap_type) end
+	if logging then
+		logit(logfile,'\n\n'..tostring(os.clock)..'(15) equip_sets: '..tostring(swap_type))
+		if val1 then
+			if val1.english then
+				logit(logfile,' : '..val1.english)
+			end
+		else
+			logit(logfile,' : nil-or-false')
+		end
+		if val2 then
+			if val2.type then	logit(logfile,' : '..val2.type)	end
+		else
+			logit(logfile,' : nil-or-false')
+		end
 	end
-	action_sent = false
-end
 
-
------------------------------------------------------------------------------------
---Name: delayed_cast(...)
---Desc: Triggers an outgoing action packet (if the passed key is valid).
---Args:
----- {...} - space delimited key for out_arr (hopefully)
------------------------------------------------------------------------------------
---Returns:
----- none
------------------------------------------------------------------------------------
-function delayed_cast(...)
-	local inde = table.concat({...},' ')
-	if out_arr[inde] then
-		local sent = out_arr[inde].data
-		windower.packets.inject_outgoing(sent:byte(2)%2*256+sent:byte(1),sent)
-	elseif debugging >= 1 or _settings.debug_mode then
-		windower.add_to_chat(8,'GearSwap (Debug Mode): Bad index passed to delayed_cast')
+	if swap_type == 'pretarget' then
+		_global.cast_delay = 0
+		if _settings.debug_mode then windower.add_to_chat(8,'GearSwap (Debug Mode): Pretarget '..tostring(val1.name)) end
+		if type(user_env.pretarget) == 'function' then user_env.pretarget(val1,val2) -- User defined function to determine the pretarget set
+		elseif user_env.pretarget then windower.add_to_chat(123,'GearSwap: pretarget() exists but is not a function') end
+		ind = mk_out_arr_entry(val1,{target_id=spell.target.id},nil)
+	elseif swap_type == 'precast' then
+		if _settings.debug_mode then windower.add_to_chat(8,'GearSwap (Debug Mode): Precast '..tostring(val1.name)) end
+		if type(user_env.precast) == 'function' then user_env.precast(val1,val2) -- User defined function to determine the precast set
+		elseif user_env.precast then windower.add_to_chat(123,'GearSwap: precast() exists but is not a function') end
+		_global.midaction = true
+	elseif swap_type == 'midcast' then
+		if _settings.debug_mode then windower.add_to_chat(8,'GearSwap (Debug Mode): Midcast '..tostring(val1.name)) end
+		user_env.midcast(val1,val2) -- User defined function to determine the midcast set
+	elseif swap_type == 'aftercast' then
+		if _settings.debug_mode then windower.add_to_chat(8,'GearSwap (Debug Mode): Aftercast '..tostring(val1.name)) end
+		if not val1 then val1 = {}
+			if debugging >= 2 then
+				windower.add_to_chat(8,'val1 error')
+			end
+		end
+		_global.midaction = false
+		user_env.aftercast(val1,val2) -- User defined function to determine the aftercast set
+	elseif swap_type == 'pet_midcast' then
+		if _settings.debug_mode then windower.add_to_chat(8,'GearSwap (Debug Mode): Pet Midcast '..tostring(val1.name)) end
+		user_env.pet_midcast(val1,val2) -- User defined function to determine the midcast set
+	elseif swap_type == 'pet_aftercast' then
+		if _settings.debug_mode then windower.add_to_chat(8,'GearSwap (Debug Mode): Pet Aftercast '..tostring(val1.name)) end
+		if val1 then
+		else
+			windower.add_to_chat(8,'---------------- VAL1 IS NIL ------------')
+		end
+		user_env.pet_aftercast(val1,val2) -- User defined function to determine the aftercast set
+	elseif swap_type == 'status_change' then
+		if _settings.debug_mode then windower.add_to_chat(8,'GearSwap (Debug Mode): Status Change '..tostring(val1)..' '..tostring(val2))
+			for i=1,#val2 do
+				windower.add_to_chat(8,'GearSwap (Debug Mode): '..i..' '..string.byte(val2[i]))
+			end
+		end
+		if type(user_env.status_change) == 'function' then user_env.status_change(val1,val2) -- User defined function to determine if sets should change following status change
+		elseif user_env.status_change then windower.add_to_chat(123,'GearSwap: status_change() exists but is not a function') end
+	elseif swap_type == 'buff_change' then
+		if _settings.debug_mode then windower.add_to_chat(8,'GearSwap (Debug Mode): Buff Change '..tostring(val1)..' '..tostring(val2)) end
+		if type(user_env.buff_change) == 'function' then user_env.buff_change(val1,val2) -- User defined function to determine if sets should change following buff change
+		elseif user_env.buff_change then windower.add_to_chat(123,'GearSwap: buff_change() exists but is not a function') end
+	elseif swap_type == 'equip_command' then
+		equip(val1)
+	elseif swap_type == 'self_command' then
+		if _settings.debug_mode then windower.add_to_chat(8,'GearSwap (Debug Mode): Self Command '..tostring(val1)) end
+		if type(user_env.self_command) == 'function' then user_env.self_command(val1)
+		elseif user_env.self_command then windower.add_to_chat(123,'GearSwap: self_command() exists but is not a function') end
+	elseif swap_type == 'delayed' then
+		equip(stored_equip_list)
+	end
+	
+	if player.race == 'Precomposed NPC' then
+		-- Short circuit the routine and get out if there's no swapping to be done because the user is a monster.
+		if swap_type == 'pretarget' then
+			return command_send_check(ind)
+		elseif swap_type == 'precast' then
+			packet_send_check(ind)
+		elseif swap_type == 'aftercast' then
+			d_out_arr_entry(val1,ind)
+		end
+		return
+	end
+	
+	
+	for i,v in pairs(short_slot_map) do
+		if equip_list[i] and (disable_table[v] or encumbrance_table[v]) then
+			not_sent_out_equip[i] = equip_list[i]
+		end
+	end
+	
+	local equip_next = to_id_set(items.inventory,equip_list) -- Translates the equip_list from the player (i=slot name, v=item name) into a table with i=slot id and v=inventory id.
+	equip_next = eliminate_redundant(cur_equip,equip_next) -- Eliminate the equip commands for items that are already equipped
+	
+	if _settings.show_swaps and table.length(equip_next)>0 then
+		local tempset = to_names_set(equip_next,items.inventory)
+		print_set(tempset,swap_type)
+	end
+	
+	local failure_reason
+	for i,v in pairs(player.buffs) do
+		if v==14 or v == 17 then
+			failure_reason = 'Charmed'
+		elseif v == 0 then
+			failure_reason = 'KOed'
+		end
+		if _settings.debug_mode and failure_reason then
+			windower.add_to_chat(8,'GearSwap (Debug Mode): Cannot change gear right now: '..failure_reason)
+		end
+	end
+	
+	if not failure_reason then
+		local one_sent
+		for _,i in ipairs(equip_order) do
+			if debugging >= 3 and equip_next[i] then
+				local out_str = 'Order: '..tostring(_)..'  Slot ID: '..tostring(i)..'  Inv. ID: '..tostring(equip_next[i])
+				if equip_next[i] ~= 0 then
+					out_str = out_str..'  Item: '..tostring(r_items[items.inventory[equip_next[i]].id][language..'_log'])
+				else
+					out_str = out_str..'  Emptying slot'
+				end
+				windower.add_to_chat(8,'GearSwap (Debugging): '..out_str)
+			elseif equip_next[i] and not disable_table[i] and not encumbrance_table[i] then
+				windower.debug('attempting to set gear. Order: '..tostring(_)..'  Slot ID: '..tostring(i)..'  Inv. ID: '..tostring(equip_next[i]))
+				windower.ffxi.set_equip(equip_next[i],i)
+				sent_out_equip[i] = equip_next[i] -- re-make the equip_next table with the name sent_out_equip as the equipment is sent out.
+			end
+		end
+	elseif logging then
+		logit(logfile,'\n\n'..tostring(os.clock)..'(69) failure_reason: '..tostring(failure_reason))
+	end
+	if debugging >= 1 then windower.debug(swap_type..' exit') end
+	
+	cache_globals(ind)
+	if swap_type == 'pretarget' then
+		command_send_check(ind)
+		if not out_arr[ind] then -- Cancelled spell
+			return
+		elseif val1.target and not st_targs:contains(val1.target.raw) then
+			equip_sets('precast',val1,val2,ind)
+			return ''
+		elseif val1.target and st_targs:contains(val1.target.raw) then
+			st_flag = true
+			return
+		end
+	elseif swap_type == 'precast' then
+		packet_send_check(ind)
+	elseif swap_type == 'aftercast' then
+		d_out_arr_entry(val1,ind)
 	end
 end
 
@@ -76,8 +240,6 @@ function cache_globals(inde)
 		end
 	end
 	_global.cast_delay = 0
-	_global.verify_equip = false
-	_global.force_send = false
 	_global.storedtarget = ''
 	_global.midaction = false
 	_global.cancel_spell = false
@@ -86,172 +248,54 @@ end
 
 
 -----------------------------------------------------------------------------------
---Name: mk_out_arr_entry(sp,arr,original)
---Desc: Makes a new entry in out_arr or updates an old one's "data" field.
+--Name: command_send_check(inde)
+--Desc: Check at the end of pretarget to see whether or not the command should be sent.
 --Args:
----- sp - Resources line for the current spell
----- arr - table containing a "target_id" field that is the spell's target_id or nil
----- original - outgoing packet string (or nil in pretarget)
+---- inde - out_arr index of the current spell
 -----------------------------------------------------------------------------------
 --Returns:
----- inde - key for out_arr
+---- string - gets propagated back to the outgoing_text function
 -----------------------------------------------------------------------------------
-function mk_out_arr_entry(sp,arr,original)
-	local inde = unify_prefix[spell.prefix]..' '..spell.english
-	if out_arr[inde..' '..tostring(arr.target_id)] then
-		inde = inde..' '..tostring(arr.target_id)
-		out_arr[inde].data = original
-		out_arr[inde].spell.target = sp.target
-	elseif out_arr[inde..' nil'] then
-		inde = inde..' nil'
-		out_arr[inde].data = original
-		out_arr[inde].spell.target = sp.target
+function command_send_check(inde)
+	if out_arr[inde].cancel_spell then
+		if debugging>=2 then windower.add_to_chat(5,'Spell canceled.') end
+		storedcommand = nil
+		out_arr[inde] = nil
 	else
-		if debugging >= 2 then windower.add_to_chat(8,'GearSwap (Debug Mode): Creating a new out_arr entry: '..tostring(inde)..' '..tostring(arr.target_id)) end
-		inde = inde..' '..tostring(arr.target_id)
-		out_arr[inde] = {}
-		out_arr[inde].data = original
-		out_arr[inde].verify_equip = false
-		out_arr[inde].cast_delay = 0
-		out_arr[inde].force_send = false
-		out_arr[inde].spell = sp
-	end
-	return inde
-end
-
-
-
------------------------------------------------------------------------------------
---Name: d_out_arr_entry(sp,ind)
---Desc: Deletes an entry from out_arr.
---Args:
----- sp - Resources line for the current spell (at the end of aftercast)
----- ind - Proposed index of out_arr
------------------------------------------------------------------------------------
---Returns:
----- None
------------------------------------------------------------------------------------
-function d_out_arr_entry(sp,ind)
-	if ind == true then -- ambiguous case
-		local deletion_table = {}
-		for i,v in pairs(out_arr) do
-			if v.midaction then
-				deletion_table[i] = true
+		out_arr[inde].spell = spell
+		if spell.target and spell.target.id and spell.prefix and unify_prefix[spell.prefix] then
+			if spell.prefix == '/item' then
+				-- Item use packet handling here
+				if spell.target.id == player.id then
+					--0x37 packet
+					out_arr[inde].proposed_packet = assemble_use_item_packet(spell.target.id,spell.target.index,spell.id)
+				else
+					--0x36 packet
+					test_packet = assemble_menu_item_packet(spell.target.id,spell.target.index,spell.id)
+					out_arr[inde].proposed_packet = test_packet
+				end
+				if not out_arr[inde].proposed_packet then
+					out_arr[inde] = nil
+				end
+			elseif outgoing_action_category_table[unify_prefix[spell.prefix]] then
+				if outgoing_action_category_table[unify_prefix[spell.prefix]] == 3 then
+					out_arr[inde].proposed_packet = assemble_action_packet(spell.target.id,spell.target.index,outgoing_action_category_table[unify_prefix[spell.prefix]],spell.index)
+				else
+					out_arr[inde].proposed_packet = assemble_action_packet(spell.target.id,spell.target.index,outgoing_action_category_table[unify_prefix[spell.prefix]],spell.id)
+				end
+			else
+				windower.add_to_chat(8,"GearSwap: Hark, what weird prefix through yonder window breaks? "..tostring(spell.prefix))
 			end
 		end
-		for i,v in pairs(deletion_table) do
-			out_arr[i] = nil
-		end
-	elseif ind == unify_prefix[sp.prefix]..' '..sp.english then
-		if out_arr[ind..' '..tostring(sp.target.id)] then
-			out_arr[ind..' '..sp.target.id] = nil
-		elseif out_arr[ind..' nil'] then
-			out_arr[ind..' nil'] = nil
-		else
-			windower.add_to_chat(123,'GearSwap: Ind identified but not found.')
-		end
-	else
-		windower.add_to_chat(123,'GearSwap: Missing ind was passed.')
-	end
-	return inde
-end
-
-
------------------------------------------------------------------------------------
---Name: out_action(arr,original)
---Desc: Determines the current spell from a given outgoing action packet.
---      Also infers whether or not it will be interrupted, which will be removed.
---Args:
----- arr - table containing outgoing action packet fields
----- original - outgoing packet string
------------------------------------------------------------------------------------
---Returns:
----- true (to block) or the outgoing packet
------------------------------------------------------------------------------------
-function out_action(arr,original)
-	spell = nil
-	local int_flag, acttype
-	if arr.category == 3 then -- 3 = Magic
-		acttype = 'Magic'
-		spell = r_spells[arr.param]
-		if buffactive.silence or buffactive.mute then int_flag = true end
-	elseif arr.category == 7 or category == 25 then -- 7 = WS, 25 = Monster skill
-		acttype = "Weapon Skill"
-		spell = r_abilities[arr.param+768]
-		if buffactive.amnesia then int_flag = true end
-	elseif arr.category == 9 then -- 9 = Ability
-		acttype = "Ability"
-		spell = r_abilities[arr.param]
-		if buffactive.amnesia then int_flag = true end
-	elseif arr.category == 16 then -- 16 = . . . ranged attack
-		acttype = "Ranged Attack"
-		spell = r_abilities[1]
-	end
-	if buffactive.terror or buffactive.sleep or buffactive.stun or buffactive.petrification or buffactive.charm then
-		int_flag = true
-	end
-	if logging then
-		local actor_name = windower.ffxi.get_mob_by_id(arr.actor_id).name
-		local target_name = windower.ffxi.get_mob_by_index(arr.target_index).name
-		logit(logfile,'\n\nActor: '..tostring(actor_name)..'  Target: '..tostring(target_name)..'  Category: '..tostring(arr.category)..'  param: '..tostring(spell.name or arr.param))
-	end
-	
-	if spell then
-		local inde = mk_out_arr_entry(spell,arr,original)
-		spell = aftercast_cost(spell)
-		if int_flag then
-			spell.interrupted = true
-		else
-			spell.interrupted = false
-		end
-		spell.name = spell[language]
-		spell.target = target_complete(windower.ffxi.get_mob_by_index(arr.target_index))
-		if _settings.debug_mode then windower.add_to_chat(8,"GearSwap (Debug Mode): Attempting to use "..spell.name) end
-		
-		return equip_sets('precast',spell,{type=acttype},inde)
 	end
 end
 
 
------------------------------------------------------------------------------------
---Name: out_item(arr,original)
---Desc: Determines the current spell from a given outgoing item use packet.
---Args:
----- arr - table containing outgoing action packet fields
----- original - outgoing packet string
------------------------------------------------------------------------------------
---Returns:
----- true (to block) or the outgoing packet
------------------------------------------------------------------------------------
-function out_item(arr,original)
-	items = windower.ffxi.get_items()
-	if arr.inventory_index == 0 then -- Gil is inventory_index 0 and item ID 65535
-		spell = aftercast_cost(r_items[65535])
-	else
-		spell = aftercast_cost(r_items[items.inventory[arr.inventory_index].id])
-	end
-	if spell then
-		local inde = mk_out_arr_entry(spell,arr,original)
-		spell.name = spell[language]
-		spell.target = target_complete(windower.ffxi.get_mob_by_index(arr.target_index))
-		if buffactive.muddle or buffactive.medicine then -- What exactly does medicated status block?
-			spell.interrupted = true
-		else
-			spell.interrupted = false
-		end
-		if _settings.debug_mode then windower.add_to_chat(8,"GearSwap (Debug Mode): Attempting to use "..item.name) end
-		return equip_sets('precast',spell,{type="Item"},inde)
-	end
-end
-
 
 -----------------------------------------------------------------------------------
---Name: packet_send_check(val,inde)
+--Name: packet_send_check(inde)
 --Desc: Determines whether or not to send the current packet.
 --      Cancels if _global.cancel_spell is true
---      Sends if val is true (_global.force_send or another)
---      Sends if verify_equip is false
---      Sends if verify_equip is true and sent_out_equip's length is 1
 --          If out_arr[inde].cast_delay is not 0, cues delayed_cast with the proper
 --          delay instead of sending immediately.
 --Args:
@@ -260,16 +304,14 @@ end
 --Returns:
 ---- true (to block) or the outgoing packet
 -----------------------------------------------------------------------------------
-function packet_send_check(val,inde)
+function packet_send_check(inde)
 	if out_arr[inde] then
 		if out_arr[inde].cancel_spell then
-			action_sent = true
 			out_arr[inde] = nil
-			sent_out_equip.ind = nil
-		elseif val or not out_arr[inde].verify_equip or ((table.length(sent_out_equip) == 1) and ( tostring(sent_out_equip.ind) ~= 'nil' ) ) then
-			action_sent = true
+		else
 			if out_arr[inde].cast_delay == 0 then
-				return out_arr[inde].data
+				send_action(inde)
+				return
 			else
 				windower.send_command('@wait '..out_arr[inde].cast_delay..';lua i '.._addon.name..' delayed_cast '..inde)
 			end
@@ -278,173 +320,84 @@ function packet_send_check(val,inde)
 	return true
 end
 
-function inc_action(act)
-	if debugging >= 1 then windower.debug('action') end
-	if gearswap_disabled or act.category == 1 then return end
-	
-	local temp_player = windower.ffxi.get_player()
-	local temp_player_mob_table = windower.ffxi.get_mob_by_index(temp_player.index)
-	local player_id = temp_player.id
-	-- Update player info for aftercast costs.
-	player.tp = temp_player.vitals.tp
-	player.mp = temp_player.vitals.mp
-	player.mpp = temp_player.vitals.mpp
-	
-	local temp_pet,pet_id
-	if temp_player_mob_table.pet_index then
-		temp_pet = windower.ffxi.get_mob_by_index(temp_player_mob_table.pet_index)
-		if temp_pet then
-			pet_id = temp_pet.id
-		end
-	end
 
-	if act.actor_id ~= player_id and act.actor_id ~= pet_id then
-		return -- If the action is not being used by the player, the pet, or is a melee attack then abort processing.
-	end
-	
-	local prefix = ''
-	
-	if act.actor_id == pet_id then 
-		prefix = 'pet_'
-	end
-	
-	spell = get_spell(act)
-	local category = act.category
-	if logging then	
-		if spell then logit(logfile,'\n\n'..tostring(os.clock)..'(178) Event Action: '..tostring(spell.english)..' '..tostring(act['category']))
-		else logit(logfile,'\n\nNil spell detected') end
-	end
-	
-	local inde
-	if spell and spell.english then
-		inde = unify_prefix[spell.prefix]..' '..spell.english
-		spell.target = target_complete(windower.ffxi.get_mob_by_id(act.targets[1].id))
-	elseif spell then
-		unknown_out_arr_deletion(prefix,{target_id = act.targets[1].id})
-		return
-	end
-	
-	if jas[category] or uses[category] or (readies[category] and act.param == 28787 and not (category == 9 or (category == 7 and prefix == 'pet_'))) then
-		-- For some reason avatar Out of Range messages send two packets (Category 4 and Category 7)
-		-- Category 4 contains real information, while Category 7 does not.
-		-- I do not know if this will affect automatons being interrupted.
-		local action_type = get_action_type(category)
-		if readies[category] and act.param == 28787 and not (category == 9) then
-			act.interrupted = true
-			action_type = 'Interruption'
-		end
-		
-		if type(user_env[prefix..'aftercast']) == 'function' then
-			equip_sets(prefix..'aftercast',spell,{type=action_type},inde)
-		elseif user_env[prefix..'aftercast'] then
-			d_out_arr_entry(spell,inde)
-			windower.add_to_chat(123,'GearSwap: '..prefix..'aftercast() exists but is not a function')
-		else
-			d_out_arr_entry(spell,inde)
-		end
-	elseif readies[category] and act.param ~= 28787 then
-		if type(user_env[prefix..'midcast']) == 'function' then
-			equip_sets(prefix..'midcast',spell,{type=get_action_type(category)},inde)
-		elseif user_env[prefix..'midcast'] then
-			windower.add_to_chat(123,'GearSwap: '..prefix..'midcast() exists but is not a function')
-		end
-	end
-end
-
-function inc_action_message(arr)
-	-- actor_id,target_id,param_1,param_2,param_3,actor_index,target_index,message_id)
-	if T{6,20,113,406,605,646}:contains(arr.message_id) then
-		-- If your current spell's target is defeated or falls to the ground
-		delete_out_arr_by_id(arr.target_id)
-	end
-	
-	local tempplay = windower.ffxi.get_player()
-	local prefix = ''
-	if arr.actor_id ~= tempplay.id then
-		if tempplay.pet_index then
-			if arr.actor_id ~= windower.ffxi.get_mob_by_index(tempplay.pet_index).id then
-				return
-			else
-				prefix = 'pet_'
-			end
-		else
-			return
-		end
-	end
-	
-	if arr.message_id == 62 then
-		if type(user_env.aftercast) == 'function' then
-			local tempitem = r_items[param_1]
-			tempitem.interrupted = true
-			equip_sets('aftercast',tempitem,{type='Interruption'},true)
-		elseif user_env.aftercast then
-			delete_out_arr_by_id(arr.target_id)
-			windower.add_to_chat(123,'GearSwap: aftercast() exists but is not a function')
-		else
-			delete_out_arr_by_id(arr.target_id)
-		end
-	elseif unable_to_use:contains(arr.message_id) then
-		if logging then	logit(logfile,'\n\n'..tostring(os.clock)..'(195) Event Action Message: '..tostring(message_id)..' Interrupt') end
-		unknown_out_arr_deletion(prefix,arr)
-	end
-end
-
-function unknown_out_arr_deletion(prefix,arr)
-	if type(user_env[prefix..'aftercast']) == 'function' then
-		local loop_check
-		for i,v in pairs(out_arr) do
-			if v.spell and v.spell.target and v.spell.target.id == arr.target_id then
-				v.spell.interrupted = true
-				equip_sets(prefix..'aftercast',v.spell,{type='Interruption'},true)
-				loop_check = true
-				break
-			end
-		end
-		if not loop_check then
-	--				equip_sets(prefix..'aftercast',{name="Unknown Interruption",english="Unknown Interruption",interrupted=true},{type='Interruption'},true)
-		end
-	elseif user_env[prefix..'aftercast'] then
-		delete_out_arr_by_id(arr.target_id)
-		windower.add_to_chat(123,'GearSwap: '..prefix..'aftercast() exists but is not a function')
-	else
-		delete_out_arr_by_id(arr.target_id)
-	end
-end
-
-function delete_out_arr_by_id(id)
-	local deleted_table = {}
-	for i,v in pairs(out_arr) do
-		if v.spell and v.spell.target then
-			if v.spell.target.id == id then
-				deleted_table[i] = true
-			end
-		end
-	end
-	for i,v in pairs(deleted_table) do
-		out_arr[i] = nil
-	end
-end
-
-function command_send_check(inde)
+-----------------------------------------------------------------------------------
+--Name: delayed_cast(...)
+--Desc: Triggers an outgoing action packet (if the passed key is valid).
+--Args:
+---- {...} - space delimited key for out_arr (hopefully)
+-----------------------------------------------------------------------------------
+--Returns:
+---- none
+-----------------------------------------------------------------------------------
+function delayed_cast(...)
+	local inde = table.concat({...},' ')
 	if out_arr[inde] then
-		if out_arr[inde].cancel_spell then
-			if debugging>=2 then windower.add_to_chat(5,'Spell canceled.') end
-			storedcommand = nil
-			return ''
-		else
-			out_arr[inde].spell = spell
-	--		print('arg1: '..tostring(val and storedcommand)..'  arg2: '..tostring(storedcommand)..' '..tostring(table.length(sent_out_equip))..'  arg3: '..tostring(storedcommand and not _global.verify_equip))
-			if (out_arr[inde].spell and out_arr[inde].spell.target and out_arr[inde].spell.target.raw) and storedcommand then
-				local assemblecommand = storedcommand..out_arr[inde].spell.target.raw
-				storedcommand = nil
-				if logging then logit(logfile,'Command Sent: '..assemblecommand..'\n') end
-				return assemblecommand
-			elseif _settings.debug_mode then
-				windower.add_to_chat(8,'GearSwap (Debug Mode): Spell was not sent because there was no stored command ('..tostring(stored_command)') or target ('..tostring(out_arr[inde].spell and out_arr[inde].spell.target and out_arr[inde].spell.target.raw)..')')
-				return ''
-			end
-		end
-	elseif debugging>=1 then
-		windower.add_to_chat(8,'GearSwap: Command send check error - inde is invalid: '..tostring(inde))
+		send_action(inde)
+	elseif debugging >= 1 or _settings.debug_mode then
+		windower.add_to_chat(8,'GearSwap (Debug Mode): Bad index passed to delayed_cast')
 	end
 end
+
+
+-----------------------------------------------------------------------------------
+--Name: send_action(inde)
+--Desc: Sends the cued action packet, if it exists.
+--Args:
+---- inde - Key to an out_arr entry that includes an action packet (hopefully)
+-----------------------------------------------------------------------------------
+--Returns:
+---- none
+-----------------------------------------------------------------------------------
+function send_action(inde)
+	if out_arr[inde].proposed_packet then
+		cued_packet = inde
+		windower.packets.inject_outgoing(out_arr[inde].proposed_packet:byte(1),out_arr[inde].proposed_packet)
+		equip_sets('midcast',out_arr[inde].spell,{type=out_arr[inde].spell.type},inde)
+		windower.send_command('input /echo Sending!;input /assist <me>')
+	else
+		windower.add_to_chat(123,'GearSwap: Cued Packet not found')
+	end
+end
+
+
+-----------------------------------------------------------------------------------
+--Name: outgoing chunk(id,original,modified,injected,blocked)
+--Desc: Searches the outgoing chunks for a packet corresponding to /assist <me>.
+--      If found, blocks that packet.
+--Args:
+---- id - ID of the current outgoing chunk
+---- original - Original outgoing chunk from the buffer
+---- modified - Outgoing chunk from the buffer after modification
+----      by other addons/plugins
+---- injected - Boolean indicating whether or not the packet was injected
+---- blocked - Boolean indicating whether or not the packet is currently blocked
+-----------------------------------------------------------------------------------
+--Returns:
+---- true if blocking the packet (/assist <me>)
+-----------------------------------------------------------------------------------
+windower.register_event('outgoing chunk',function(id,original,modified,injected,blocked)
+	if gearswap_disabled then return end
+	if debugging >= 1 then windower.debug('outgoing chunk '..id) end
+	if id == 0x1A and not injected then
+		local cur_time = os.clock()
+		for i,v in pairs(outgoing_packet_table) do
+			if cur_time-v > 1 then
+				outgoing_packet_table[i] = nil
+			elseif i:sub(1,2) == original:sub(1,2) and i:sub(5) == original:sub(5) then
+				return
+			end
+		end
+		outgoing_packet_table[original] = os.clock()
+
+		local target_index = get_bit_packed(original,64,80)
+		local category = get_bit_packed(original,80,96)
+		local target_id = windower.ffxi.get_mob_by_index(target_index).id
+		if category == 12 and cued_packet and out_arr[cued_packet] and out_arr[cued_packet].proposed_packet and target_id == player.id then
+			cued_packet = nil
+			return true
+		end
+	elseif id == 0x1A and injected and clocking then
+		windower.add_to_chat(8,'Injection time: '..(os.clock()-out_time))
+	end
+end)
