+++ addons/GearSwap/statics.lua
@@ -167,13 +167,13 @@ uses = {false,true,true,true,true,false,false,false,false,false,true,false,true,
 unable_to_use = T{17,18,55,56,87,88,89,90,104,191,308,313,325,410,428,561,574,579,580,581,661,665,
 	4,5,12,16,34,35,40,47,48,49,71,72,76,78,84,91,92,94,95,96,106,111,128,154,155,190,192,193,198,
 	199,215,216,217,218,219,220,233,246,247,307,315,316,328,337,338,346,347,348,349,356,411,443,444,
-	445,446,514,516,517,518,523,524,525,547,568,569,575,649,660,662,666,700,701} -- Probably don't need some of these (event action)
+	445,446,514,516,517,518,523,524,525,547,568,569,575,649,660,662,666,700,701,62} -- Probably don't need some of these (event action)
 
 -- 192 : param_1 = Ability ID
 -- 17 : no information
 -- 34 : param_1 = Spell index
 pass_through_targs = T{'<t>','<me>','<ft>','<scan>','<bt>','<lastst>','<r>','<pet>','<p0>','<p1>','<p2>','<p3>','<p4>',
-	'<p5>','<a10>','<a11>','<a12>','<a13>','<a14>','<a15>','<a20>','<a21>','<a22>','<a23>','<a24>','<a25>','<stnpc>',
+	'<p5>','<a10>','<a11>','<a12>','<a13>','<a14>','<a15>','<a20>','<a21>','<a22>','<a23>','<a24>','<a25>','<st>','<stnpc>',
 	'<stal>','<stpc>','<stpt>'}
 avatar_element = {Ifrit='Fire',Titan='Earth',Leviathan='Water',Garuda='Wind',Shiva='Ice',Ramuh='Thunder',Carbuncle='Light',
 	Diabolos='Dark',Fenrir='Dark',['Fire Elemental']='Fire',['Earth Elemental']='Earth',['Water Elemental']='Water',
@@ -252,21 +252,10 @@ short_slot_map.right_ring = 14
 short_slot_map.back = 15
 
 
--- key mapping for combining sets
-function defaultSlotMap(k)
-	if slot_map[k] then
-		return default_slot_map[slot_map[k]]
-	else
-		return nil
-	end
-end
-
-
 
 _global = make_user_table()
 _global.cast_delay = 0
 _global.storedtarget = ''
-_global.verify_equip = false
 _global.cancel_spell = false
 _global.midaction = false
 _global.current_event = 'None'
@@ -276,6 +265,7 @@ _settings.debug_mode = false
 _settings.show_swaps = false
 
 
+last_PC_update = ''
 gearswap_disabled = false
 sent_out_equip = {}
 not_sent_out_equip = {}
@@ -283,7 +273,6 @@ limbo_equip = {}
 out_arr = {}
 equip_list = {}
 equip_order = {}
-action_sent = false
 world = make_user_table()
 buffactive = make_user_table()
 player = make_user_table()
@@ -293,9 +282,10 @@ pet = make_user_table()
 pet.isvalid = false
 fellow = make_user_table()
 fellow.isvalid = false
-st_flag = false
+st_targs = T{'<st>','<stpc>','<stal>','<stnpc>','<stpt>'}
 current_job_file = nil
 disable_table = {false,false,false,false,false,false,false,false,false,false,false,false,false,false,false}
+outgoing_action_category_table = {['/ma']=3,['/ws']=7,['/ja']=9,['/ra']=16,['/ms']=25}
 disable_table[0] = false
 encumbrance_table = table.reassign({},disable_table)
 registered_user_events = {}
